732846 - adding a javascript lint to our unit tests and fixing errors

+++ run-tests.sh
@@ -1,4 +1,14 @@
 #!/bin/bash
+
+# Check JavaScript syntax
+jsl --conf scripts/jsl-errors-only.conf 
+OUT=$?
+if [ $OUT -eq 3 ];then
+  echo "SYNTAX ERROR detected in JavaScript files in src/public/javascripts/"
+  exit 1
+fi
+
+# Run unit tests
 cd src
 sudo bundle install
 rake db:migrate:reset --trace
